Move Image and IconAction to own componenent files
